+++ internal.go
@@ -1,0 +1,7 @@
+package mysqlx
+
+var internal = struct {
+	debugf func(string, ...any)
+}{
+	debugf: func(s string, a ...any) {},
+}