+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Supports MySQL `OR`.
 - Now prototypes parameter could either be a struct or a pointer to struct.
 - Cancel go 1.1 in CI test.
-- Defaultly use `charset=utfmb4` instead of `utf8`.
+- Use `charset=utf8mb4` instead of `utf8` by default.
 
 ## v0.3.0